Fixed serialization issue with Charity.last_year_donation_total.cents

+++ everydayhero.Api/v3/Entities/Cause.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace everydayhero.Api.v3
 {
     public class Cause
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string name { get; set; }
         public string key { get; set; }
     }
